(PC-21481) chore(fetchOffers): rename fetchOffer into fetchOffers to be more accurate
